fix for L8 authenticates sessions middleware
